% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\docType{data}
\name{DR_DM}
\alias{DR_DM}
\title{Simple features for the "Municipal Districts of the Dominican Republic"}
\format{
\subsection{\code{DR_DM}}{

A simple features collection with 393 features and 2 fields:
\describe{
\item{DM_ID}{Municipal district number (6 digits) based on the \verb{División Territorial 2021} order}
\item{geometry}{Polygon geometry representing each municipal district}
}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_DM}
}
\usage{
DR_DM
}
\description{
This dataset contains simple features polygons representing the 393 municipal districts
of the Dominican Republic, as defined in the \verb{División Territorial 2021}.
}
\keyword{datasets}
