% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\docType{data}
\name{DR_MUN}
\alias{DR_MUN}
\title{Municipalities of the Dominican Republic}
\format{
\subsection{\code{DR_MUN}}{

A simple features collection with 158 features and 2 fields:
\describe{
\item{MUN_ID}{Municipality number (3 digits) based on the \verb{División Territorial 2021} order}
\item{geometry}{Polygon geometry representing each municipality}
}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_MUNICIPIOS}
}
\usage{
DR_MUN
}
\description{
This dataset contains simple features polygons representing the 158 municipalities
of the Dominican Republic, as defined in the \verb{División Territorial 2021}.
}
\keyword{datasets}
