% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\docType{data}
\name{dr_municipal_district}
\alias{dr_municipal_district}
\title{Municipal Districts of the Dominican Republic}
\format{
\subsection{\code{dr_municipal_district}}{

A data frame with 393 rows and 3 variables:
\describe{
\item{DM_ID}{Municipal district number (6 digits) based on the \verb{División Territorial 2021} order}
\item{DM_NAME}{Full name of the municipal district as defined in the \verb{División Territorial 2021}}
}
}
}
\source{
\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_municipal_district
}
\description{
This dataset contains the metadata for the 393 municipal districts of the Dominican
Republic as defined in the \verb{División Territorial 2021}. Due to historical name changes,
some municipal districts may have duplicate IDs. Users should be aware of this when
analyzing the data.
}
\keyword{datasets}
