% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dr_municipal_districts}
\alias{dr_municipal_districts}
\title{Load Dominican Republic Municipal Districts Simple Features (sf)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
A simple features collection with 393 features and 10 fields:
\describe{
\item{MD_ID}{Municipal district number (6 digits) based on the \verb{División Territorial 2021} order}
\item{MD_NAME}{Full name of the municipal district as defined in the \verb{División Territorial 2021}}
\item{MUN_ID}{Municipality number (4 digits)}
\item{MUN_NAME}{Full name of the municipality}
\item{PROV_ID}{Province number (2 digits)}
\item{PROV_CODE}{Code of the province}
\item{PROV_NAME}{Full name of the province}
\item{REG_ID}{Region number}
\item{REG_NAME}{Name of the region}
\item{REG_CODE}{Code of the region}
\item{geometry}{Polygon geometry representing each municipal district}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_DM}

\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_municipal_districts(
  .sf = TRUE,
  .mun = FALSE,
  .prov = FALSE,
  .reg = FALSE,
  .uniques = TRUE
)
}
\arguments{
\item{.sf}{Logical; if TRUE, converts the result to an \code{sf} object. Defaults to TRUE.}

\item{.mun}{Logical; if FALSE, removes columns related to municipalities. Defaults to FALSE.}

\item{.prov}{Logical; if FALSE, removes columns related to provinces. Defaults to FALSE.}

\item{.reg}{Logical; if FALSE, removes columns related to regions. Defaults to FALSE.}

\item{.uniques}{Logical; if TRUE, removes some duplicated IDs inserted to consider
territories names changes over time. Defaults to TRUE.}
}
\value{
A data frame or \code{sf} object (if .sf is TRUE) with the loaded and processed data.
}
\description{
This function loads the Dominican Republic municipal districts simple features (sf)
and adds metadata for municipalities, provinces, and regions if required.
It also optionally converts the result to an \code{sf} object.
}
\examples{
\dontrun{
# Load and process the data as sf object
result <- dr_municipal_districts(TRUE, TRUE, TRUE, TRUE)
print(result)
}

}
