% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{dr_municipalities}
\alias{dr_municipalities}
\title{Load Dominican Republic Municipalities Simple Features (sf)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
A simple features collection with 158 features and 8 fields:
\describe{
\item{MUN_ID}{Municipality number (3 digits) based on the \verb{División Territorial 2021} order}
\item{MUN_NAME}{Full name of the municipality as defined in the \verb{División Territorial 2021}}
\item{PROV_ID}{Province number (2 digits)}
\item{PROV_CODE}{Code of the province}
\item{PROV_NAME}{Full name of the province}
\item{REG_ID}{Region number}
\item{REG_NAME}{Name of the region}
\item{REG_CODE}{Code of the region}
\item{geometry}{Polygon geometry representing each municipality}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_MUNICIPIOS}

\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_municipalities(.sf = TRUE, .prov = FALSE, .reg = FALSE)
}
\arguments{
\item{.sf}{Logical; if TRUE, converts the result to an \code{sf} object. Defaults to TRUE.}

\item{.prov}{Logical; if FALSE, removes columns related to provinces. Defaults to FALSE.}

\item{.reg}{Logical; if FALSE, removes columns related to regions. Defaults to FALSE.}
}
\value{
A data frame or \code{sf} object (if .sf is TRUE) with the loaded and processed data.
}
\description{
This function loads the Dominican Republic municipalities simple features (sf)
and adds metadata for provinces and regions if required.
It also optionally converts the result to an \code{sf} object.
}
\examples{
\dontrun{
  # Load and process the data as sf object
  result <- dr_municipalities(TRUE, TRUE, TRUE)
  print(result)
}

}
