% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\docType{data}
\name{dr_municipality}
\alias{dr_municipality}
\title{Municipalities of the Dominican Republic}
\format{
\subsection{\code{dr_municipality}}{

A data frame with 158 rows and 3 variables:
\describe{
\item{MUN_ID}{Municipality number (3 digits) based on the \verb{División Territorial 2021} order}
\item{MUN_NAME}{Full name of the municipality as defined in the \verb{División Territorial 2021}}
}
}
}
\source{
\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_municipality
}
\description{
This dataset contains the metadata for the 158 municipalities of the Dominican
Republic as defined in the \verb{División Territorial 2021}. Note that due to historical
name changes, some municipalities may have duplicate IDs. Users should consider this
when working with the data.
}
\keyword{datasets}
