% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\docType{data}
\name{DR_PROV}
\alias{DR_PROV}
\title{Provinces of the Dominican Republic}
\format{
\subsection{\code{DR_PROV}}{

A simple features collection with 32 features and 2 fields:
\describe{
\item{PROV_ID}{Province number (2 digits) based on the \verb{División Territorial 2021} order}
\item{geometry}{Polygon geometry representing each province}
}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_PROV}
}
\usage{
DR_PROV
}
\description{
This dataset contains simple features polygons representing the 32 provinces
of the Dominican Republic, as defined in the \verb{División Territorial 2021}.
}
\keyword{datasets}
