% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\docType{data}
\name{dr_province}
\alias{dr_province}
\title{Provinces of the Dominican Republic}
\format{
\subsection{\code{dr_province}}{

A data frame with 32 rows and 3 variables:
\describe{
\item{PROV_ID}{Province number (2 digits) based on the \verb{División Territorial 2021} order}
\item{PROV_NAME}{Full name of the province as defined in the \verb{División Territorial 2021}}
\item{PROV_CODE}{Province code, a custom 2 or 3-letter abbreviation created for each province}
}
}
}
\source{
\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_province
}
\description{
This dataset contains the metadata for the 32 provinces of the Dominican
Republic as defined in the \verb{División Territorial 2021}. Note that some
provinces have undergone name changes over time, leading to potential
duplicates in province IDs. Users should be aware of this when analyzing
the data.
}
\keyword{datasets}
