% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{dr_provinces}
\alias{dr_provinces}
\title{Load Dominican Republic Provinces Simple Features (sf)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
A simple features collection with 32 features and 6 fields:
\describe{
\item{PROV_ID}{Province number (2 digits) based on the \verb{División Territorial 2021} order}
\item{PROV_CODE}{Province code, a custom 2 or 3-letter abbreviation created for each province}
\item{PROV_NAME}{Full name of the province as defined in the \verb{División Territorial 2021}}
\item{REG_ID}{Region number}
\item{REG_NAME}{Name of the region}
\item{REG_CODE}{Code of the region}
\item{geometry}{Polygon geometry representing each province}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_PROV}

\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_provinces(.sf = TRUE, .reg = FALSE)
}
\arguments{
\item{.sf}{Logical; if TRUE, converts the result to an \code{sf} object. Defaults to TRUE.}

\item{.reg}{Logical; if FALSE, removes columns related to regions. Defaults to FALSE.}
}
\value{
A data frame or \code{sf} object (if .sf is TRUE) with the loaded and processed data.
}
\description{
This function loads the Dominican Republic provinces simple features (sf)
and adds metadata for regions if required.
It also optionally converts the result to an \code{sf} object.
}
\examples{
\dontrun{
  # Load and process the data as sf object
  result <- dr_provinces(TRUE, TRUE)
  print(result)
}

}
