% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\docType{data}
\name{DR_REG}
\alias{DR_REG}
\title{Development Regions of the Dominican Republic}
\format{
\subsection{\code{DR_REG}}{

A simple features collection with 10 features and 2 fields:
\describe{
\item{REG_ID}{Region number (2 digits) based on the \verb{Decreto 710-04} order}
\item{geometry}{Polygon geometry}
}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_REG_20220630}
}
\usage{
DR_REG
}
\description{
This dataset contains the simple features polygons for the development
regions of the Dominican Republic. These regions correspond to the 10
development regions as defined in \verb{Decreto 710-04}.
}
\keyword{datasets}
