% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\docType{data}
\name{dr_region}
\alias{dr_region}
\title{Metadata for the Development Regions of the Dominican Republic}
\format{
\subsection{\code{dr_region}}{

A data frame with 10 rows and 3 variables:
\describe{
\item{REG_ID}{Region number (2 digits) based on the \verb{Decreto 710-04} order}
\item{REG_NAME}{Full name of the region as defined in the \verb{Decreto 710-04}}
\item{REG_CODE}{Region code (3 letters). Abbreviation of the full region name}
}
}
}
\source{
\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_region
}
\description{
This dataset contains the metadata for the development regions of the
Dominican Republic. The regions correspond to the 10 development regions
defined in \verb{Decreto 710-04}.
}
\keyword{datasets}
