% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{dr_regions}
\alias{dr_regions}
\title{Load Dominican Republic Development Regions Simple Features (sf)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
A simple features collection with 10 features and 3 fields:
\describe{
\item{REG_ID}{Region number (2 digits) based on the \verb{Decreto 710-04} order}
\item{REG_NAME}{Full name of the region as defined in the \verb{Decreto 710-04}}
\item{REG_CODE}{Region code (3 letters). Abbreviation of the full region name}
\item{geometry}{Polygon geometry representing each region}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_REG_20220630}

\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_regions(.sf = TRUE)
}
\arguments{
\item{.sf}{Logical; if TRUE, converts the result to an \code{sf} object. Defaults to TRUE.}
}
\value{
A data frame or \code{sf} object (if .sf is TRUE) with the loaded and processed data.
}
\description{
This function loads the Dominican Republic development regions simple features (sf)
and adds metadata. It also optionally converts the result to an \code{sf} object.
}
\examples{
\dontrun{
  # Load and process the data as sf object
  result <- dr_regions()
  print(result)
}

}
