% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\docType{data}
\name{DR_SEC}
\alias{DR_SEC}
\title{Simple features for the "Sections of the Dominican Republic"}
\format{
\subsection{\code{DR_SEC}}{

A simple features collection with 1,565 features and 2 fields:
\describe{
\item{SEC_ID}{Section number (8 digits) based on the \verb{División Territorial 2021} order}
\item{geometry}{Polygon geometry representing each section}
}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_SECCIONES}
}
\usage{
DR_SEC
}
\description{
This dataset contains simple features polygons representing the 1,565 sections
of the Dominican Republic, as defined in the \verb{División Territorial 2021}.
}
\keyword{datasets}
