% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\docType{data}
\name{dr_section}
\alias{dr_section}
\title{Sections of the Dominican Republic}
\format{
\subsection{\code{dr_section}}{

A data frame with 1,565 rows and 3 variables:
\describe{
\item{SEC_ID}{Section number (8 digits) based on the \verb{División Territorial 2021} order}
\item{SEC_NAME}{Full name of the section as defined in the \verb{División Territorial 2021}}
}
}
}
\source{
\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_section
}
\description{
This dataset contains the metadata for the 1,565 sections of the Dominican
Republic as defined in the \verb{División Territorial 2021}. Due to historical name changes,
some sections may have duplicate IDs. Users should consider this when working with the data.
}
\keyword{datasets}
