% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\name{dr_sections}
\alias{dr_sections}
\title{Load Dominican Republic Sections Simple Features (sf)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\format{
A simple features collection with 1,565 features and 12 fields:
\describe{
\item{SEC_ID}{Section number (8 digits) based on the \verb{División Territorial 2021} order}
\item{SEC_NAME}{Full name of the section as defined in the \verb{División Territorial 2021}}
\item{MD_ID}{Municipal district number (6 digits)}
\item{MD_NAME}{Full name of the municipal district}
\item{MUN_ID}{Municipality number (4 digits)}
\item{MUN_NAME}{Full name of the municipality}
\item{PROV_ID}{Province number (2 digits)}
\item{PROV_CODE}{Code of the province}
\item{PROV_NAME}{Full name of the province}
\item{REG_ID}{Region number}
\item{REG_NAME}{Name of the region}
\item{REG_CODE}{Code of the region}
\item{geometry}{Polygon geometry representing each section}
}
}
\source{
\url{https://geoportal.iderd.gob.do/layers/geonode_data:geonode:RD_SECCIONES}

\url{https://www.one.gob.do/publicaciones/2021/division-territorial-2021/}
}
\usage{
dr_sections(.sf = TRUE, .dm = FALSE, .mun = FALSE, .prov = FALSE, .reg = FALSE)
}
\arguments{
\item{.sf}{Logical; if TRUE, converts the result to an \code{sf} object. Defaults to TRUE.}

\item{.dm}{Logical; if FALSE, removes columns related to municipal districts. Defaults to FALSE.}

\item{.mun}{Logical; if FALSE, removes columns related to municipalities. Defaults to FALSE.}

\item{.prov}{Logical; if FALSE, removes columns related to provinces. Defaults to FALSE.}

\item{.reg}{Logical; if FALSE, removes columns related to regions. Defaults to FALSE.}
}
\value{
A data frame or \code{sf} object (if .sf is TRUE) with the loaded and processed data.
}
\description{
This function loads the Dominican Republic sections simple features (sf)
and adds metadata for municipal districts, municipalities, provinces, and regions if required.
It also optionally converts the result to an \code{sf} object.
}
\examples{
\dontrun{
  # Load and process the data as sf object
  result <- dr_sections(TRUE, TRUE, TRUE, TRUE, TRUE)
  print(result)
}

}
