% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{get_dr_provinces}
\alias{get_dr_provinces}
\title{Deprecated function: get_dr_provinces
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
get_dr_provinces(...)
}
\arguments{
\item{...}{Arguments to pass to \code{dr_provinces()}.}
}
\description{
This function is deprecated. Use \code{dr_provinces()} instead.
}
