% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remesas_provincias}
\alias{remesas_provincias}
\title{Percentage distribution by provinces of remittances received by the Dominican Republic}
\format{
\subsection{\code{remesas_provincias}}{

\describe{
\item{province}{Full name of the province}
\item{percentage}{Percentage of the total remittances received in 2010}
}
}
}
\source{
\url{https://www.bancentral.gov.do/a/d/2532-sector-externo}
}
\usage{
remesas_provincias
}
\description{
Percentage distribution by provinces of remittances received by the Dominican Republic
}
\keyword{datasets}
