% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remittance_by_province}
\alias{remittance_by_province}
\title{Percentage distribution by provinces of remittances received by the Dominican Republic}
\format{
\subsection{\code{remittance_by_province}}{

\describe{
\item{province}{Full name of the province}
\item{percentage}{Percentage of the total remittances received in 2010}
}
}
}
\source{
\url{https://www.bancentral.gov.do/a/d/2532-sector-externo}
}
\usage{
remittance_by_province
}
\description{
Percentage distribution by provinces of remittances received by the Dominican Republic
}
\keyword{datasets}
