% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfDR-package.R
\docType{package}
\name{sfDR-package}
\alias{sfDR}
\alias{sfDR-package}
\title{sfDR: Simple Features for Dominican Republic Administrative Boundaries}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The sfDR package provides a comprehensive collection of simple feature (sf) objects for the administrative boundaries of the Dominican Republic, including provinces, municipalities, municipal districts, and sections. These datasets are ideal for geospatial analysis, mapping, and data visualization, providing accurate and up-to-date information as per the `División Territorial 2021`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adatar-do.github.io/sfDR/}
}

}
\author{
\strong{Maintainer}: Daniel E. de la Rosa \email{me@dnldelarosa.com}

Other contributors:
\itemize{
  \item Adatar [copyright holder]
}

}
\keyword{internal}
