% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{get_dr_provinces}
\alias{get_dr_provinces}
\title{Deprecated function: get_dr_provinces}
\usage{
get_dr_provinces(...)
}
\arguments{
\item{...}{Arguments to pass to \code{dr_provinces()}.}
}
\description{
This function is deprecated. Use \code{dr_provinces()} instead.
}
